+++ src/utils/Contact.java
@@ -1,17 +1,17 @@
 package utils;
 
 public class Contact {
-	private int id;
+	private long id;
 	private String name;
 	private String email;
 	
-	public Contact(int id, String name, String email) {
+	public Contact(long id, String name, String email) {
 		this.id = id;
 		this.name = name;
 		this.email = email;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
